fix rnn to train on sequences
This is likely the reason for poor performance. It has no history
because the RNN wasn't working, it was just passing through the z_dim as
best it could. This is because I was giving it sequences of 1 with a
detached hidden state, so it treated the hidden state and a fixed input
variable. Bugger.